+++ core/lib.go
@@ -1,21 +1,5 @@
 package core
 
-import (
-	"fmt"
-	color "github.com/logrusorgru/aurora"
-)
-
-func check(e error) {
-	if e != nil {
-		panic(e)
-	}
-}
-
-func PrintCommand(command *Command) {
-	fmt.Println(color.Bold("Command:"))
-	fmt.Println(command.Command)
-}
-
 func GetAllTags(projects []Project) map[string]struct{} {
 	tags := make(map[string]struct{})
 	for _, project := range projects {